requirements.txt
numpy~=1.26.3
torchaudio~=2.1.2
torch~=2.1.2
tqdm~=4.66.1
transformers~=4.36.2
matplotlib~=3.8.2
gradio~=4.15.0
pyannote.core
datasets~=2.16.1
accelerate~=0.25.0
GitPython~=3.1.41